# Docs/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Swagger_RentMoto"},"host":"localhost:63177","schemes":["http"],"paths":{"/api/entregadores":{"post":{"tags":["entregadores"],"summary":"Adicionar Entregador no sistema","operationId":"entregadores_Add","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"entregador","in":"body","required":true,"schema":{"$ref":"#/definitions/Entregador"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/entregadores/{id}":{"get":{"tags":["entregadores"],"summary":"Busca Entregador pelo ID de cadastro","operationId":"entregadores_GetById","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"post":{"tags":["entregadores"],"summary":"Enviar arquivo de imagem da CNH - Somente PNG e BMP","operationId":"entregadores_UploadCNH","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"delete":{"tags":["entregadores"],"summary":"Remover um entregador","operationId":"entregadores_Delete","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/locacao":{"get":{"tags":["locacao"],"summary":"Calcula o valor da Locação com dados inputados. Ele não busca ou persiste dados, ele realiza o cálculo de acordo com o json Locacao enviado","operationId":"locacao_GetContractValue","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"locacao.id","in":"query","required":false,"type":"integer","format":"int32"},{"name":"locacao.entregador_id","in":"query","required":true,"type":"string"},{"name":"locacao.moto_id","in":"query","required":true,"type":"string"},{"name":"locacao.data_inicio","in":"query","required":true,"type":"string","format":"date-time"},{"name":"locacao.data_termino","in":"query","required":true,"type":"string","format":"date-time"},{"name":"locacao.data_previsao_termino","in":"query","required":true,"type":"string","format":"date-time"},{"name":"locacao.plano","in":"query","required":true,"type":"integer","format":"int32"},{"name":"locacao.data_devolucao","in":"query","required":false,"type":"string","format":"date-time"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"post":{"tags":["locacao"],"summary":"Adicionar uma Locação com a Moto e o Entregador informados","operationId":"locacao_Add","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"locacao","in":"body","required":true,"schema":{"$ref":"#/definitions/Locacao"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/locacao/{id}":{"get":{"tags":["locacao"],"summary":"Buscar Locacao por ID","operationId":"locacao_GetById","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"put":{"tags":["locacao"],"summary":"Informar uma Data de Devolução","operationId":"locacao_Devolucao","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"locacao","in":"body","required":true,"schema":{"$ref":"#/definitions/Locacao"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/motos":{"get":{"tags":["motos"],"summary":"Retorna uma Lista com Todas as Motos do Sistema","operationId":"motos_GetAllMoto","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"post":{"tags":["motos"],"summary":"Insere uma Moto no sistema","operationId":"motos_CreateMoto","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"moto","in":"body","required":true,"schema":{"$ref":"#/definitions/Moto"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/motos/{id}/placa":{"get":{"tags":["motos"],"summary":"Busca uma Moto pela Placa","operationId":"motos_GetMotoByPlaca","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"placa","in":"query","required":false,"type":"string"},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/motos/{id}":{"get":{"tags":["motos"],"summary":"Busca uma Moto no sistema pelo ID Identidade","operationId":"motos_GetMotoById","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"put":{"tags":["motos"],"summary":"Atualizar a placa de uma moto já cadastrada no sistema","operationId":"motos_UpdatePlacaMoto","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"moto","in":"body","required":true,"schema":{"$ref":"#/definitions/Moto"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"delete":{"tags":["motos"],"summary":"Apagar uma Moto já gravada no sistema","operationId":"motos_DeleteMoto","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}}},"definitions":{"Entregador":{"required":["identificador","nome","cnpj","numero_cnh","tipo_cnh"],"type":"object","properties":{"id":{"format":"int32","type":"integer"},"identificador":{"type":"string"},"nome":{"type":"string"},"cnpj":{"type":"string"},"data_nascimento":{"format":"date-time","type":"string"},"numero_cnh":{"type":"string"},"tipo_cnh":{"pattern":"^(A|B|AB)$","type":"string"},"imagem_cnh":{"type":"string"}}},"Locacao":{"required":["entregador_id","moto_id","data_inicio","data_termino","data_previsao_termino","plano"],"type":"object","properties":{"id":{"format":"int32","type":"integer"},"entregador_id":{"type":"string"},"moto_id":{"type":"string"},"data_inicio":{"format":"date-time","type":"string"},"data_termino":{"format":"date-time","type":"string"},"data_previsao_termino":{"format":"date-time","type":"string"},"plano":{"format":"int32","type":"integer"},"data_devolucao":{"format":"date-time","type":"string"},"valor_diaria":{"format":"double","type":"number","readOnly":true}}},"Moto":{"required":["identificador","modelo","ano"],"type":"object","properties":{"identificador":{"type":"string"},"modelo":{"type":"string"},"ano":{"format":"int32","type":"integer"},"placa":{"type":"string"}}}}}